refactor: consolidate checker into single file
Merge all check_*.c/h files and checker_util.c/h back into checker.c/h,
mirroring the parser consolidation in #24.
