--- w0/check_decl.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_CHECK_DECL_H
-#define WHIST_CHECK_DECL_H
-
-#include "ast.h"
-#include "checker.h"
-
-void check_decl(Checker* checker, Node* node);
-
-#endif
