--- w0/check_expression.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_CHECK_EXPRESSION_H
-#define WHIST_CHECK_EXPRESSION_H
-
-#include "ast.h"
-#include "checker.h"
-
-Type* check_expression(Checker* checker, Node* node);
-
-#endif
