--- w0/check_statement.h
@@ -1,9 +1,0 @@
-#ifndef WHIST_CHECK_STATEMENT_H
-#define WHIST_CHECK_STATEMENT_H
-
-#include "ast.h"
-#include "checker.h"
-
-void check_statement(Checker* checker, Node* node);
-
-#endif
